retry decorator refactoring, downloading extra GFS timestep for potential edge case
